Avoid listings (FastCGI); OpenBSD requires no socket
We have to dance a little to avoid leaking unnecessary bits,
as the FastCGI module automatically list directories. Note
that tyring to access /data will send an index.html, which will
fail on the front (not a SPA page).

Also, OpenBSD requires the socket not to exist at all before
starting. Haven't digged deeper.